route config: moving the route provider manager to its own file
