+++ source/common/router/BUILD
@@ -177,8 +177,14 @@ envoy_cc_library(
 
 envoy_cc_library(
     name = "rds_lib",
-    srcs = ["rds_impl.cc"],
-    hdrs = ["rds_impl.h"],
+    srcs = [
+        "rds_impl.cc",
+        "route_provider_manager.cc",
+    ],
+    hdrs = [
+        "rds_impl.h",
+        "route_provider_manager.h",
+    ],
     deps = [
         ":config_lib",
         "//envoy/config:subscription_interface",
