+++ source/extensions/filters/network/http_connection_manager/config.cc
@@ -33,7 +33,7 @@
 #include "source/common/local_reply/local_reply.h"
 #include "source/common/protobuf/utility.h"
 #include "source/common/quic/server_connection_factory.h"
-#include "source/common/router/rds_impl.h"
+#include "source/common/router/route_provider_manager.h"
 #include "source/common/runtime/runtime_impl.h"
 #include "source/common/tracing/custom_tag_impl.h"
 #include "source/common/tracing/tracer_config_impl.h"
@@ -533,11 +533,16 @@ HttpConnectionManagerConfig::HttpConnectionManagerConfig(
   switch (config.route_specifier_case()) {
   case envoy::extensions::filters::network::http_connection_manager::v3::HttpConnectionManager::
       RouteSpecifierCase::kRds:
+    route_config_provider_ = route_config_provider_manager.createRdsRouteConfigProvider(
+        // At the creation of a RDS route config provider, the factory_context's initManager is
+        // always valid, though the init manager may go away later when the listener goes away.
+        config.rds(), context_.serverFactoryContext(), stats_prefix_, context_.initManager());
+    break;
   case envoy::extensions::filters::network::http_connection_manager::v3::HttpConnectionManager::
       RouteSpecifierCase::kRouteConfig:
-    route_config_provider_ = Router::RouteConfigProviderUtil::create(
-        config, context_.serverFactoryContext(), context_.messageValidationVisitor(),
-        context_.initManager(), stats_prefix_, route_config_provider_manager_);
+    route_config_provider_ = route_config_provider_manager.createStaticRouteConfigProvider(
+        config.route_config(), context_.serverFactoryContext(),
+        context_.messageValidationVisitor());
     break;
   case envoy::extensions::filters::network::http_connection_manager::v3::HttpConnectionManager::
       RouteSpecifierCase::kScopedRoutes:
